Revert "E shop"
