+++ .env
@@ -1,3 +1,3 @@
-DATABASE_URL=postgres://wlla@localhost:5432/wm
+DATABASE_URL=postgres://hgjiyvhprjcmae:99ecf0d10f78993080f468bd0726347502fcc514c445ee92a7207b757cdc5f3a@ec2-52-210-97-223.eu-west-1.compute.amazonaws.com:5432/d2pgji158gbv7l
 JWT_SECRET=hamzah
-PORT=4000
+PORT=4001
