--- models/comment.model.js
@@ -1,12 +1,0 @@
-"use strict";
-
-function createCommentTable(sequelize, DataTypes) {
-  return (
-    sequelize.define("comment", {
-      userId: { type: DataTypes.INTEGER, allowNull: false },
-      memoryId: { type: DataTypes.INTEGER, allowNull: false },
-      comment: { type: DataTypes.STRING, allowNull: false }     
-    })
-  )
-}
-module.exports = { createCommentTable };
