+++ package.json
@@ -20,14 +20,12 @@
   },
   "homepage": "https://github.com/School-trip-app/School-trip-app-backEnd#readme",
   "dependencies": {
-    "axios": "^1.1.3",
     "base-64": "^1.0.0",
     "bcrypt": "^5.1.0",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "jsonwebtoken": "^8.5.1",
-    "multer": "^1.4.5-lts.1",
     "node": "^18.10.0",
     "nodemon": "^2.0.20",
     "pg": "^8.8.0",
